requirements.txt
# ── Dados e ML ────────────────────────────────────────────────────────────
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
imbalanced-learn>=0.11

# ── Gradient Boosting ──────────────────────────────────────────────────────
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2

# ── Explicabilidade ────────────────────────────────────────────────────────
shap>=0.44

# ── Visualização ──────────────────────────────────────────────────────────
matplotlib>=3.7
seaborn>=0.12

# ── API ───────────────────────────────────────────────────────────────────
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.0

# ── Utilitários ───────────────────────────────────────────────────────────
joblib>=1.3
python-dotenv>=1.0
requests>=2.31

# ── Notebooks ─────────────────────────────────────────────────────────────
jupyter>=1.0
ipykernel>=6.0
ipywidgets>=8.0

# ── Testes ────────────────────────────────────────────────────────────────
pytest>=7.4
pytest-cov>=4.1
